switch_rtp windows tweaks and type changes.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1026 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/endpoints/mod_dingaling/mod_dingaling.c
@@ -131,7 +131,7 @@ struct private_object {
 	unsigned char out_digit_packet[4];
 	unsigned int out_digit_sofar;
 	unsigned int out_digit_dur;
-	unsigned int out_digit_seq;
+	uint16_t out_digit_seq;
 	int32_t timestamp_send;
 	int32_t timestamp_recv;
 	int32_t timestamp_dtmf;
